+++ libfm/libfm/learn_sgd_element_adapt_reg.h
@@ -76,7 +76,7 @@ protected:
                 var_v(f) += fm->v(f,j)*fm->v(f,j);
             }
         }
-        mean_w /= (double) fm->num_attribute;
+        mean_w /= (float) fm->num_attribute;
         var_w = var_w/fm->num_attribute - mean_w*mean_w;
         for (int f = 0; f < fm->num_factor; f++) {
             mean_v(f) /= fm->num_attribute;
@@ -90,8 +90,8 @@ protected:
     }
 
     void sgd_theta_step(SparseRow<float>& x, const float target) {
-        double p = fm->predict(x, sum, sum_sqr);
-        double mult = 0;
+        float p = fm->predict(x, sum, sum_sqr);
+        float mult = 0;
         if (task == TASK_REGRESSION) {
             p = std::fminf(fm->max_target, p);
             p = std::fmaxf(fm->min_target, p);
@@ -102,14 +102,14 @@ protected:
 
         // make the update with my regularization constants:
         if (fm->is_use_w0) {
-            double& w0 = fm->w0;
-            double grad_0 = mult;
+            float& w0 = fm->w0;
+            float grad_0 = mult;
             w0 -= learn_rate * (grad_0 + 2 * reg_0 * w0);
         }
         if (fm->is_use_w) {
             for (uint i = 0; i < x.size; i++) {
                 uint g = x.col2group[x.ids[i]];
-                double& w = fm->w(x.ids[i]);
+                float& w = fm->w(x.ids[i]);
                 grad_w(x.ids[i]) = mult * x.values[i];
                 w -= learn_rate * (grad_w(x.ids[i]) + 2 * reg_w(g) * w);
             }
@@ -117,15 +117,15 @@ protected:
         for (int f = 0; f < fm->num_factor; f++) {
             for (uint i = 0; i < x.size; i++) {
                 uint g = x.col2group[x.ids[i]];
-                double& v = fm->v(f,x.ids[i]);
+                float& v = fm->v(f,x.ids[i]);
                 grad_v(f,x.ids[i]) = mult * (x.values[i] * (sum(f) - v * x.values[i])); // grad_v_if = (y(x)-y) * [ x_i*(\sum_j x_j v_jf) - v_if*x^2 ]
                 v -= learn_rate * (grad_v(f,x.ids[i]) + 2 * reg_v(g,f) * v);
             }
         }
     }
 
-    double predict_scaled(SparseRow<float>& x) {
-        double p = 0.0;
+    float predict_scaled(SparseRow<float>& x) {
+        float p = 0.0;
         if (fm->is_use_w0) {
             p += fm->w0;
         }
@@ -133,8 +133,8 @@ protected:
             for (uint i = 0; i < x.size; i++) {
                 assert(x.ids[i] < fm->num_attribute);
                 uint g = x.col2group[x.ids[i]];
-                double& w = fm->w(x.ids[i]);
-                double w_dash = w - learn_rate * (grad_w(x.ids[i]) + 2 * reg_w(g) * w);
+                float& w = fm->w(x.ids[i]);
+                float w_dash = w - learn_rate * (grad_w(x.ids[i]) + 2 * reg_w(g) * w);
                 p += w_dash * x.values[i];
             }
         }
@@ -143,9 +143,9 @@ protected:
             sum_sqr(f) = 0.0;
             for (uint i = 0; i < x.size; i++) {
                 uint g = x.col2group[x.ids[i]];
-                double& v = fm->v(f,x.ids[i]);
-                double v_dash = v - learn_rate * (grad_v(f,x.ids[i]) + 2 * reg_v(g,f) * v);
-                double d = v_dash * x.values[i];
+                float& v = fm->v(f,x.ids[i]);
+                float v_dash = v - learn_rate * (grad_v(f,x.ids[i]) + 2 * reg_v(g,f) * v);
+                float d = v_dash * x.values[i];
                 sum(f) += d;
                 sum_sqr(f) += d*d;
             }
@@ -155,8 +155,8 @@ protected:
     }
 
     void sgd_lambda_step(SparseRow<float>& x, const float target) {
-        double p = predict_scaled(x);
-        double grad_loss = 0;
+        float p = predict_scaled(x);
+        float grad_loss = 0;
         if (task == 0) {
             p = std::fminf(fm->max_target, p);
             p = std::fmaxf(fm->min_target, p);
@@ -182,21 +182,21 @@ protected:
             // sum_f_dash      := \sum_{l} x_l * v'_lf, this is independent of the groups
             // sum_f(g)        := \sum_{l \in group(g)} x_l * v_lf
             // sum_f_dash_f(g) := \sum_{l \in group(g)} x^2_l * v_lf * v'_lf
-            double sum_f_dash = 0.0;
+            float sum_f_dash = 0.0;
             sum_f.init(0.0);
             sum_f_dash_f.init(0.0);
             for (uint i = 0; i < x.size; i++) {
                 // v_if' =  [ v_if * (1-alpha*lambda_v_f) - alpha * grad_v_if]
                 uint g = x.col2group[x.ids[i]];
-                double& v = fm->v(f,x.ids[i]);
-                double v_dash = v - learn_rate * (grad_v(f,x.ids[i]) + 2 * reg_v(g,f) * v);
+                float& v = fm->v(f,x.ids[i]);
+                float v_dash = v - learn_rate * (grad_v(f,x.ids[i]) + 2 * reg_v(g,f) * v);
 
                 sum_f_dash += v_dash * x.values[i];
                 sum_f(g) += v * x.values[i];
                 sum_f_dash_f(g) += v_dash * x.values[i] * v * x.values[i];
             }
             for (uint g = 0; g < fm->num_group; g++) {
-                double lambda_v_grad = -2 * learn_rate *  (sum_f_dash * sum_f(g) - sum_f_dash_f(g));
+                float lambda_v_grad = -2 * learn_rate *  (sum_f_dash * sum_f(g) - sum_f_dash_f(g));
                 reg_v(g,f) -= learn_rate * grad_loss * lambda_v_grad;
                 reg_v(g,f) = std::fmaxf(0.0, reg_v(g,f));
             }
